Align SeedPrompt with PromptRequestPiece (#1053)
