+++ doc/code/memory/8_seed_prompt_database.py
@@ -67,8 +67,9 @@ for prompt in prompts:
 
 # %% [markdown]
 # ## Adding multimodal Seed Prompt Groups to the database
-# In this next example, we will add a Seed Prompt group with prompts across the audio, image, video, and text modalities.
-# The Seed Prompt groups have the same `prompt_group_id`, meaning they will be sent together.
+# In this next example, we will add a Seed Prompt Group with prompts across the audio, image, video, and text modalities.
+# Seed Prompts that have the same `prompt_group_alias` will be part of the same Seed Prompt Group. Within a Seed Prompt Group,
+# Seed Prompts that share a `sequence` will be sent together as part of the same turn (e.g. text and corresponding image).
 # When we add non-text seed prompts to memory, encoding data will automatically populate in the seed prompt's
 # `metadata` field, including `format` (i.e. png, mp4, wav, etc.) as well as additional metadata for audio
 # and video files, inclduing `bitrate` (kBits/s as int), `samplerate` (samples/second as int), `bitdepth` (as int),
@@ -84,9 +85,6 @@ from pyrit.models import SeedPromptGroup
 seed_prompt_group = SeedPromptGroup.from_yaml_file(
     pathlib.Path(DATASETS_PATH) / "seed_prompts" / "illegal-multimodal-group.prompt"
 )
-
-# Render user-defined values for yaml template
-seed_prompt_group.render_template_value(stolen_item="a car")
 
 await memory.add_seed_prompt_groups_to_memory(prompt_groups=[seed_prompt_group], added_by="test multimodal illegal")  # type: ignore
