+++ tests/unit/test_normalizer_request.py
@@ -6,7 +6,8 @@ import uuid
 
 import pytest
 
-from pyrit.models.seed_prompt import SeedPrompt, SeedPromptGroup
+from pyrit.models.seed_prompt import SeedPrompt
+from pyrit.models.seed_prompt_group import SeedPromptGroup
 from pyrit.prompt_normalizer import NormalizerRequest
 
 
@@ -15,8 +16,8 @@ def test_normalizer_request_validates_sequence():
 
     seed_prompt_group = SeedPromptGroup(
         prompts=[
-            SeedPrompt(value="Hello", data_type="text", prompt_group_id=group_id, sequence=0),
-            SeedPrompt(value="World", data_type="text", prompt_group_id=group_id, sequence=1),
+            SeedPrompt(value="Hello", data_type="text", prompt_group_id=group_id, sequence=0, role="user"),
+            SeedPrompt(value="World", data_type="text", prompt_group_id=group_id, sequence=1, role="user"),
         ]
     )
